backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-learning-backend"
version = "0.1.0"
description = "智能化论文学习平台后端骨架"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.15.2",
  "celery[redis]>=5.4.0",
  "fastapi>=0.110.0",
  "oss2>=2.19.1",
  "psycopg[binary]>=3.1.18",
  "pydantic-settings>=2.2.1",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.28",
  "uvicorn[standard]>=0.28.0"
]

[tool.setuptools.packages.find]
include = ["app*"]
